fix : 에러 확인

+++ src/main/java/com/example/vblogserver/domain/review/controller/ReviewController.java
@@ -164,7 +164,7 @@ public class ReviewController {
         System.out.println("9");
         String reviewContent = requestReviewDTO.getContent();
         System.out.println("10");
-        float grade = requestReviewDTO.getGrade();
+        float grade = (float)requestReviewDTO.getGrade();
         System.out.println("11");
         //LoginID 로 userID 조회
         User user;
